+++ examples/run_finetune.py
@@ -330,7 +330,7 @@ def main():
             logger.info("Benchmark done.")
         else:
             if not training_args.autotuner_benchmark:
-                trainer.save_model(merge_tensor_parallel=training_args.tensor_parallel_degree > 1)
+                trainer.save_model(merge_tensor_parallel=training_args.tensor_parallel_degree > 1, fc_to_hf=True)
                 trainer.log_metrics("train", train_result.metrics)
                 trainer.save_metrics("train", train_result.metrics)
                 trainer.save_state()
