+++ paddleformers/nn/lm_head.py
@@ -14,9 +14,11 @@
 
 import paddle
 import paddle.nn as nn
-from paddle.distributed.flex_checkpoint.dcp.sharded_weight import (
-    build_sharded_state_dict,
-)
+
+try:
+    from paddle.distributed.flex_checkpoint.dcp.sharded_weight import build_sharded_state_dict
+except:
+    build_sharded_state_dict = None
 
 from ..generation.configuration_utils import PretrainedConfig
 from .criterion.loss_utils import calc_lm_head_logits
