+++ paddleformers/trainer/training_args.py
@@ -410,6 +410,10 @@ class TrainingArguments:
             Specifies the format for loading checkpoints. Options are: None, 'sharding_io', 'unified_checkpoint', 'flex_checkpoint'. (default: None). This setting is ignored if the corresponding switch is configured.
         aoa_config (`Optional[dict[str, list[str]]]`, *optional*):
             The AoA configuration of FlexCheckpoint, used to describe the mapping between model weights and the checkpoint content. Default is None.
+        load_via_cpu (bool, optional):
+            Whether to load checkpoint data into CPU memory first before transferring to GPU.
+            This helps mitigate GPU memory shortage by staging data on the CPU and only moving required parts to the GPU on demand during communication.
+            Defaults to False.
     """
 
     output_dir: str = field(
@@ -1153,6 +1157,13 @@ class TrainingArguments:
         default=None,
         metadata={
             "help": "The AoA configuration of FlexCheckpoint, used to describe the mapping between model weights and the checkpoint content. Default is None."
+        },
+    )
+
+    load_via_cpu: Optional[bool] = field(
+        default=False,
+        metadata={
+            "help": "If True, loads checkpoint data to CPU first, then transfers required parts to GPU on demand to reduce GPU memory usage. Defaults to False."
         },
     )
 
@@ -2535,8 +2546,6 @@ class TrainingArguments:
                 return True
             elif self.enable_auto_parallel:
                 return True
-            elif self.save_checkpoint_format == "flex_checkpoint":
-                return False
             elif self.use_hybrid_parallel:
                 # save on dataset rank 0
                 return (
